test(config): the guard half moves to its own file, and two more pins
tests/test_config_read_once.py reached 1123 lines, which is over the
repo's 800-line file limit for a file this lane ADDS. Split by size, at
the seam the file already had: tests/test_config_guard.py is the static
half (layers A and B, the call shapes, the disclosed limit) and
test_config_read_once.py is the behaviour half. Each says what the other
covers.

Two pins added, both because a round-2 fix would otherwise be caught by
nothing.

TestTheEnvelopeDoesNotOutliveItsRun drives run_factory twice on one
FactoryConfig with `max_files_changed` edited between them and asserts
run 2 enforces the edited value and that `policy_config` is still None.
That is the reviewer's own disproof of the round-1 "M5 is an equivalent
mutant" claim, turned into a test: with the write restored, run 2
inherits run 1's clamped envelope.

TestTheLadderStateIsReadOnce counts AutonomyState.load up to pipeline
construction and pins it at 1. No parse count covers it, because the
ladder state is JSON rather than kstrl.toml, so the duplicate read
survived a review by being nine lines from the first. The epilogue's
read is a separate question and is asserted separately:
_record_autonomy_outcome re-reads AFTER the run because apply_demotion
may have written the file while components ran.
